# Reduce allocations for pipelines using pre-allocated slice backing memory