@nussjustin 2021-01-23
@mediocregopher WDYT?

The main reason I did this is so that it's less likely that one would need to reuse or even pool `Pipeline`s because of all the allocations. While pooling is easy, it's also easy to get wrong (e.g. introduce races) and annoying. So if we can make it so that most users don't need to think about it, that seems like a win to me.

Also as explained in the issue, the array length 4 was chosen to avoid bloating memory usage to much. I also considered a length of 8 which would avoid up to 3 more allocations. Compared to 4 this would only add 320 bytes to each `pipeline`. I hadn't measured this before so I went with 4 for the initial change, but seeing that we are talking about only 320 bytes I think using 8-element arrays would be better.