@mediocregopher 2021-01-23
Hey @nussjustin ! You're way better at these optimizations than me so I'm glad you're back at it :)

I tried to design Pipeline so that it could be re-used, if the user cares to do that, but as long as the API doesn't have to change then I don't mind doing these optimizations also. Using arrays to pre-allocate the slices is a great trick that I really need to remember. And yeah I think 8 seems like a better number to me if it's just 320 bytes we're talking about, I know in my usages of Pipeline I go over 4 pretty often, but I'd guess I'm usually around or under 8.

WDYT of the pipeline implementation in general? It's a bit janky but it has the benefit of not relying directly on the `cmdAction` type.