@nussjustin 2021-01-23
Increased the array size to 8. PTAL

> WDYT of the pipeline implementation in general? It's a bit janky but it has the benefit of not relying directly on the cmdAction type.

Not having (to think about) `CmdAction` just feels so good. Also as a user of radix I really like the visible changes, especially around pipelining.

While the new code (everything, not just the pipelining) feels more complex to me, that's really just because I was quite familiar with the old code.