@mediocregopher 2021-02-08
Ah, sorry! Will merge now, thanks! I'm glad you like the new interface, it does feel a lot cleaner than before. And yeah the pipelining code in particular is pretty weird, but it works :crossed_fingers: